Add openDayBookTransaction modal
