+++ src/components/dashboard/daybook/DayBookIndex/DeleteDayBookTransaction.tsx
@@ -1,37 +1,23 @@
 import { LoadingButton } from '@mui/lab';
-import {
-  Alert,
-  Button,
-  Card,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  Stack,
-} from '@mui/material';
-import { DataGrid, GridColDef } from '@mui/x-data-grid';
+import { Button } from '@mui/material';
 import { useSnackbar } from 'notistack';
-import { FC, useMemo } from 'react';
+import { FC } from 'react';
 import { useDeleteDayBookTransaction } from 'src/hooks/cache/dayBook';
-import {
-  DayBookTableEntry,
-  DayBookTransaction,
-  DayBookTransactionDetail,
-} from 'src/types/dayBook';
+import { DayBookTransaction } from 'src/types/dayBook';
+import { OpenDayBookTransaction } from './OpenDayBookTransaction';
 
 interface DeleteDayBookTransactionProps {
   transaction: DayBookTransaction | null;
-  setTransaction: (transaction: DayBookTransaction | null) => void;
+  onClose: () => void;
 }
 
 export const DeleteDayBookTransaction: FC<DeleteDayBookTransactionProps> = ({
   transaction,
-  setTransaction,
+  onClose,
 }) => {
   const { enqueueSnackbar } = useSnackbar();
   const { mutateAsync: handleDeleteAccount, isPending } =
     useDeleteDayBookTransaction();
-  const handleClose = () => setTransaction(null);
 
   const transactionId = transaction?.id || '';
 
@@ -39,7 +25,7 @@ export const DeleteDayBookTransaction: FC<DeleteDayBookTransactionProps> = ({
     handleDeleteAccount(transactionId)
       .then(() => {
         enqueueSnackbar('Transacción eliminada exitosamente');
-        handleClose();
+        onClose();
       })
       .catch((error) => {
         console.error(error);
@@ -49,123 +35,31 @@ export const DeleteDayBookTransaction: FC<DeleteDayBookTransactionProps> = ({
       });
   };
 
-  const columns: GridColDef<DayBookTransactionDetail>[] = [
-    {
-      field: 'account_id',
-      headerName: 'Cuenta contable',
-      type: 'string',
-      flex: 3,
-    },
-    {
-      field: 'debit',
-      headerName: 'Debe',
-      headerAlign: 'center',
-      type: 'number',
-      flex: 2,
-      align: 'center',
-      valueFormatter: ({ value }) =>
-        value ? `$${Number(value).toFixed(2)}` : '-',
-    },
-    {
-      field: 'credit',
-      headerName: 'Haber',
-      headerAlign: 'center',
-      type: 'number',
-      flex: 2,
-      align: 'center',
-      valueFormatter: ({ value }) =>
-        value ? `$${Number(value).toFixed(2)}` : '-',
-    },
-    {
-      field: 'description',
-      headerName: 'Descripción',
-      type: 'string',
-      flex: 6,
-    },
-    {
-      field: 'quotation_id',
-      headerName: 'Cotización No.',
-      headerAlign: 'center',
-      type: 'number',
-      flex: 2,
-      align: 'center',
-      valueFormatter: ({ value }) => value || '-',
-    },
-    {
-      field: 'invoice_id',
-      headerName: 'Factura No.',
-      headerAlign: 'center',
-      type: 'number',
-      flex: 2,
-      align: 'center',
-      valueFormatter: ({ value }) => value || '-',
-    },
-  ];
-
-  const rows: DayBookTableEntry[] = useMemo(
-    () =>
-      transaction?.transactions.map((t) => ({
-        ...t,
-        id: t.account_id,
-        date: transaction.date,
-      })) || [],
-
-    [transaction?.date, transaction?.transactions]
-  );
-
   if (!transaction) return null;
 
   return (
-    <Dialog
-      fullWidth
-      maxWidth="lg"
-      onClose={handleClose}
-      open={Boolean(transaction)}
-    >
-      <DialogTitle>Eliminar Transacción</DialogTitle>
+    <OpenDayBookTransaction
+      transaction={transaction}
+      onClose={onClose}
+      title="Eliminar Transacción"
+      actions={
+        <>
+          <Button onClick={onClose} disabled={isPending}>
+            Cancelar
+          </Button>
 
-      <Stack component={DialogContent} gap={2}>
-        <Alert severity="warning">
-          La transacción que deseas eliminar contiene los siguientes detalles.
-          ¿Está seguro que desea eliminar esta transacción?
-        </Alert>
-
-        <Card variant="outlined">
-          <DataGrid
-            columns={columns}
-            rows={rows}
-            density="compact"
-            initialState={{
-              sorting: {
-                sortModel: [{ field: 'date', sort: 'asc' }],
-              },
-            }}
-            sx={{
-              '& .MuiDataGrid-columnHeader': {
-                bgcolor: (theme) => theme.palette.action.selected,
-              },
-            }}
-            autoHeight
-            disableRowSelectionOnClick
-            hideFooter
-          />
-        </Card>
-      </Stack>
-
-      <DialogActions>
-        <Button onClick={handleClose} disabled={isPending}>
-          Cancelar
-        </Button>
-
-        <LoadingButton
-          variant="contained"
-          onClick={handleConfirm}
-          loading={isPending}
-          color="error"
-        >
-          Eliminar
-        </LoadingButton>
-      </DialogActions>
-    </Dialog>
+          <LoadingButton
+            variant="contained"
+            onClick={handleConfirm}
+            loading={isPending}
+            color="error"
+          >
+            Eliminar
+          </LoadingButton>
+        </>
+      }
+      alertText="La transacción que deseas eliminar contiene los siguientes detalles. ¿Está seguro que desea eliminar esta transacción?"
+      alertSeverity="warning"
+    />
   );
 };
